backend/requirements.txt
fastapi[standard]
uvicorn[standard]
python-dotenv
psycopg2-binary
bcrypt